requirements.txt
discord.py-self==1.9.2
aiohttp==3.8.5
